Ported to use Maven
Given this project was VERY old, I had not learned about the wonders of
build systems quite yet. That being said, the dependencies and resources
for this project were not very organized or coherently named. This would
mean that anyone try to build this project would have a very hard time
doing so. Fortunately however, I ported it to use maven. If anyone has
issues building this project now, please feel free to open an issue in
the GitHub.